#2146 correct  to cast attribute as double, not unsigned. Same in
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@16770 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ cell_based/src/mesh/PottsMesh.cpp
@@ -451,7 +451,7 @@ void PottsMesh<DIM>::ConstructFromMeshReader(AbstractMeshReader<DIM, DIM>& rMesh
         if (rMeshReader.GetNumElementAttributes() > 0)
         {
             assert(rMeshReader.GetNumElementAttributes() == 1);
-            unsigned attribute_value = element_data.AttributeValue;
+            double attribute_value = element_data.AttributeValue;
             p_element->SetAttribute(attribute_value);
         }
     }
